// src/schedules/model/IAvConfiguration.interface.ts
export interface IAvConfiguration {
  url: string;
  firstRoute: number;
  lastRoute: number;

  xPathRoutes: string;
  xPathPlatforms: string;
  xPathRegularities: string;
  xPathDepartmentAS: string;
  xPathDepartmentTU: string;
  xPathArrival: string;
  xPathDepartmentBack: string;
  xPathArrivalBack: string;
}